move demo and home views to pages. replace package-lock.json with yarn.lock
