+++ examples/vue-example-app/src/pages/Demo.vue
