+++ examples/vue-example-app/src/pages/Home.vue
